move the metadata extraction function to the converters
