+++ safe_rcm/product/converters.py
@@ -1,3 +1,4 @@
+import datatree
 import toolz
 import xarray as xr
 
@@ -49,3 +50,32 @@ def convert_table(table, *, namespaces={}, index_hint="first", dtypes={}):
         )
         .pipe(lambda obj: obj if list(obj) != ["$"] else obj["$"].rename(None))
     )
+
+
+def metadata_filter(item, ignore):
+    """filter metadata items
+
+    Metadata items are either attributes (the name starts with an '@'), or they are scalars.
+    """
+    k, v = item
+
+    return (k.startswith("@") or utils.is_scalar(v)) and k not in ignore
+
+
+def extract_metadata(
+    mapping,
+    collapse=(),
+    ignore=("@xmlns", "@xmlns:rcm", "@xmlns:xsi", "@xsi:schemaLocation"),
+):
+    # extract the metadata
+    filter_ = toolz.functoolz.flip(metadata_filter, ignore)
+    metadata = toolz.dicttoolz.keymap(
+        lambda k: k.lstrip("@"), toolz.dicttoolz.itemfilter(filter_, mapping)
+    )
+
+    # collapse the selected items
+    to_collapse = toolz.dicttoolz.keyfilter(lambda x: x in collapse, mapping)
+    collapsed = dict(toolz.itertoolz.concat(v.items() for v in to_collapse.values()))
+
+    attrs = metadata | collapsed
+    return datatree.DataTree(xr.Dataset(attrs=attrs))
